+++ L4D2AddonAssistant.Core/WorkshopVpkAddon.cs
@@ -212,14 +212,16 @@ namespace L4D2AddonAssistant
                         Log.Error(ex, "Exception occurred during reading meta info file: {FilePath}", metaInfoPath);
                     }
                 }
+                string? vpkPathPreview = null;
                 if (metaInfo != null)
                 {
-                    var vpkPathPreview = Path.Join(dirPath, metaInfo.CurrentFile);
-                    if (File.Exists(vpkPathPreview))
+                    var path = Path.Join(dirPath, metaInfo.CurrentFile);
+                    if (File.Exists(path))
                     {
-                        _vpkPath = vpkPathPreview;
+                        vpkPathPreview = path;
                     }
                 }
+                SetVpkPath(vpkPathPreview);
 
                 _checkTask = Task.Run(() =>
                 {
@@ -323,7 +325,7 @@ namespace L4D2AddonAssistant
                     {
                         AddProblem(problem);
                     }
-                    _vpkPath = resultVpkPath;
+                    SetVpkPath(resultVpkPath);
 
                     if (task.Exception != null)
                     {
@@ -463,6 +465,12 @@ namespace L4D2AddonAssistant
                 return Task.FromResult(new GetPublishedFileDetailsResult(null, GetPublishedFileDetailsResultStatus.Failed));
             }
             return WebUtils.GetPublishedFileDetailsAsync(_publishedFileId.Value, Root.HttpClient, cancellationToken);
+        }
+
+        private void SetVpkPath(string? path)
+        {
+            _vpkPath = path;
+            NotifyChanged(nameof(FullVpkFilePath));
         }
     }
 }